requirements.txt
http://download.pytorch.org/whl/torch-0.2.0.post3-cp27-none-macosx_10_7_x86_64.whl 
torchvision 
tqdm
scikit-learn
pandas
